Factor out codegen.py.
